Use proper gcc size_t type

+++ common/source/util/SWEBDebugInfo.cpp
@@ -104,9 +104,9 @@ void SWEBDebugInfo::printCallInformation(pointer address) const {
     ssize_t line;
     getCallNameAndLine(address, name, line);
     if (line >= 0) {
-        kprintfd("%10zx: %." CALL_FUNC_NAME_LIMIT_STR "s:%zu \n", address, name, line );
+      kprintfd("%10zx: %." CALL_FUNC_NAME_LIMIT_STR "s:%zu \n", address, name, (size_t)line );
     }
     else {
-        kprintfd("%10zx: %." CALL_FUNC_NAME_LIMIT_STR "s+%zx\n", address, name, -line);
+      kprintfd("%10zx: %." CALL_FUNC_NAME_LIMIT_STR "s+%zx\n", address, name, (size_t)-line);
     }
 }
